+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 * Detect light background and adapt the text and controls to it.
+* Allow to disable the bottom bar overlay via the `bottomOverlay` prop.
 
 ## 0.1.0 (October 10, 2016)